WIP: Migrate openflowplugim-impl tests to uint types
// FIXME: finish this up

JIRA: OPNFLWPLUG-1099

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/connection/listener/OpenflowProtocolListenerInitialImplTest.java
@@ -5,7 +5,6 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.impl.connection.listener;
 
 import org.junit.Assert;
@@ -23,6 +22,7 @@ import org.opendaylight.openflowplugin.api.openflow.connection.HandshakeContext;
 import org.opendaylight.openflowplugin.api.openflow.md.core.HandshakeManager;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.EchoRequestMessageBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.HelloMessageBuilder;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 /**
  * Test for {@link OpenflowProtocolListenerInitialImpl}.
@@ -55,7 +55,7 @@ public class OpenflowProtocolListenerInitialImplTest {
     @Test
     public void testOnEchoRequestMessage() {
         EchoRequestMessageBuilder echoRequestMessageBld = new EchoRequestMessageBuilder()
-                .setXid(42L)
+                .setXid(Uint32.valueOf(42))
                 .setVersion(OFConstants.OFP_VERSION_1_3);
         openflowProtocolListenerInitial.onEchoRequestMessage(echoRequestMessageBld.build());
 
@@ -65,7 +65,7 @@ public class OpenflowProtocolListenerInitialImplTest {
     @Test
     public void testOnHelloMessage() {
         HelloMessageBuilder helloMessageBld = new HelloMessageBuilder()
-                .setXid(42L)
+                .setXid(Uint32.valueOf(42))
                 .setVersion(OFConstants.OFP_VERSION_1_3);
         openflowProtocolListenerInitial.onHelloMessage(helloMessageBld.build());
