+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/statistics/services/OpendaylightGroupStatisticsServiceImplTest.java
@@ -5,14 +5,12 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.impl.statistics.services;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 
 import com.google.common.util.concurrent.FutureCallback;
-import java.math.BigInteger;
 import java.util.Collections;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicLong;
@@ -55,7 +53,9 @@ import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.multipart.reply.multipart.reply.body.multipart.reply.group.features._case.MultipartReplyGroupFeaturesBuilder;
 import org.opendaylight.yangtools.yang.common.RpcResult;
 import org.opendaylight.yangtools.yang.common.RpcResultBuilder;
+import org.opendaylight.yangtools.yang.common.Uint16;
 import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 /**
  * Test for {@link OpendaylightGroupStatisticsServiceImpl}.
@@ -63,8 +63,8 @@ import org.opendaylight.yangtools.yang.common.Uint32;
 public class OpendaylightGroupStatisticsServiceImplTest extends AbstractSingleStatsServiceTest {
 
     private static final org.opendaylight.yang.gen.v1.urn
-            .opendaylight.openflow.common.types.rev130731.GroupId GROUP_ID =
-            new org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.GroupId(123L);
+            .opendaylight.openflow.common.types.rev130731.GroupId GROUP_ID = new org.opendaylight.yang.gen.v1.urn
+                .opendaylight.openflow.common.types.rev130731.GroupId(Uint32.valueOf(123));
     @Captor
     private ArgumentCaptor<MultipartRequestInput> requestInput;
 
@@ -114,9 +114,9 @@ public class OpendaylightGroupStatisticsServiceImplTest extends AbstractSingleSt
                                         .setGroupDesc(Collections.singletonList(new GroupDescBuilder()
                                                 .setGroupId(GROUP_ID)
                                                 .setBucketsList(Collections.singletonList(new BucketsListBuilder()
-                                                        .setWatchGroup(51L)
-                                                        .setWatchPort(new PortNumber(52L))
-                                                        .setWeight(53)
+                                                        .setWatchGroup(Uint32.valueOf(51))
+                                                        .setWatchPort(new PortNumber(Uint32.valueOf(52)))
+                                                        .setWeight(Uint16.valueOf(53))
                                                         .build()))
                                                 .setType(GroupType.OFPGTALL)
                                                 .build()))
@@ -168,7 +168,7 @@ public class OpendaylightGroupStatisticsServiceImplTest extends AbstractSingleSt
     public void testGetGroupStatistics() throws Exception {
         GetGroupStatisticsInputBuilder input = new GetGroupStatisticsInputBuilder()
                 .setNode(createNodeRef("unitProt:123"))
-                .setGroupId(new GroupId(21L));
+                .setGroupId(new GroupId(Uint32.valueOf(21)));
 
         rpcResult = buildGroupStatsResponse();
 
@@ -188,15 +188,15 @@ public class OpendaylightGroupStatisticsServiceImplTest extends AbstractSingleSt
                         .setMultipartReplyBody(new MultipartReplyGroupCaseBuilder()
                                 .setMultipartReplyGroup(new MultipartReplyGroupBuilder()
                                         .setGroupStats(Collections.singletonList(new GroupStatsBuilder()
-                                                .setByteCount(BigInteger.valueOf(21L))
-                                                .setPacketCount(BigInteger.valueOf(22L))
-                                                .setRefCount(23L)
-                                                .setDurationSec(24L)
-                                                .setDurationNsec(25L)
+                                                .setByteCount(Uint64.valueOf(21))
+                                                .setPacketCount(Uint64.valueOf(22))
+                                                .setRefCount(Uint32.valueOf(23))
+                                                .setDurationSec(Uint32.valueOf(24))
+                                                .setDurationNsec(Uint32.valueOf(25))
                                                 .setGroupId(GROUP_ID)
                                                 .setBucketStats(Collections.singletonList(new BucketStatsBuilder()
-                                                        .setByteCount(BigInteger.valueOf(26L))
-                                                        .setPacketCount(BigInteger.valueOf(27L))
+                                                        .setByteCount(Uint64.valueOf(26))
+                                                        .setPacketCount(Uint64.valueOf(27))
                                                         .build()))
                                                 .build()))
                                         .build())
